fix: change chapterOrder parameter type to int in existsByModuleIdAndChapterOrder method

+++ src/main/java/edu/vinu/repository/ChapterRepository.java
@@ -23,5 +23,5 @@ public interface ChapterRepository extends JpaRepository<ChapterEntity,Long> {
 
     int countByModuleId(Long moduleId);
 
-    boolean existsByModuleIdAndChapterOrder(Long moduleId, Long chapterOrder);
+    boolean existsByModuleIdAndChapterOrder(Long moduleId, int chapterOrder);
 }
